FIX completed controller delete padel match

+++ Backend/src/api/controllers/padelMatches.js
@@ -1,5 +1,10 @@
+const { idAndRoleChecked } = require("../../utils/checkId&Role");
+const { deleteImage } = require("../../utils/deleteImage");
 const { ParamsErrorOfPadelMatch } = require("../../utils/ParamsErrorOfPadelMatch");
+const { resultPadelMatchDeleted } = require("../../utils/resultPadelMatchDeleted");
+const { resultPadelMatchesByDay } = require("../../utils/resultPadelMatchesByDate");
 const PadelMatch = require("../models/padelMatches");
+const User = require("../models/users");
 
 const createPadelMatch = async (req, res, next) => {
     try {
@@ -10,10 +15,14 @@ const createPadelMatch = async (req, res, next) => {
             return res.status(400).json({ message: padelMatchParamsError });
         }
 
-        // const padelMatchDuplicated = await PadelMatch.findOne({ day, month, hour })
-        // padelMatchDuplicated ? res.status(400).json({ message: `Ya existe un partido el dia ${day} a las ${hour} creado por ${author}.` }) : padelMatchDuplicated
+        const authorId = await User.findById(req.user);
 
-        const newPadelMatch = new PadelMatch(req.body);
+        const newPadelMatch = new PadelMatch({ ...req.body, author: authorId._id });
+
+        if (req.file) {
+            newPadelMatch.image = req.file.path;
+        }
+
         const padelMatchSaved = await newPadelMatch.save();
         return res.status(201).json({ message: "Partido creado.", padelMatchSaved });
     } catch (error) {
@@ -24,27 +33,33 @@ const createPadelMatch = async (req, res, next) => {
 const getPadelMatches = async (req, res, next) => {
     try {
         const allPadelMatches = await PadelMatch.find();
-        return allPadelMatches.length
-            ? res.status(200).json({ message: `Estos son todos los partidos que hay: ${allPadelMatches}` })
+        return allPadelMatches
+            ? res.status(200).json({ message: "Estos son todos los partidos que hay programados:", allPadelMatches })
             : res.status(400).json({ message: "No hay ningún partido programado." });
     } catch (error) {
         return res.status(400).json(`❌ Fallo en getPadelMatches: ${error.message}`);
     }
 };
 
-const getPadelMatchByDate = async (req, res, next) => {
+const getPadelMatchByDay = async (req, res, next) => {
     try {
-        const { date } = req.params;
-        const findPadelMatch = await PadelMatch.find({
-            date: new RegExp(date, "i")
-        });
-        findPadelMatch.length
-            ? res.status(200).json({ message: `Se han encontrado los siguientes partidos: `, findPadelMatch })
-            : res.status(400).json({ message: `No se ha encontrado ningún partido.` });
+        const { day } = req.params;
+        const findPadelMatch = await PadelMatch.find({ day });
+        resultPadelMatchesByDay(res, findPadelMatch, day);
     } catch (error) {
-        return res.status(400).json(`❌ Fallo en getPadelMatchByDate: ${error.message}`);
+        return res.status(400).json(`❌ Fallo en getPadelMatchByDay: ${error.message}`);
     }
 };
+
+// const getPadelMatchByAuthor = async (req, res, next) => {
+//     try {
+//         const { author } = req.params;
+//         const findPadelMatch = await PadelMatch.find({ author });
+//         resultPadelMatchesByAuthor(res, findPadelMatch, author);
+//     } catch (error) {
+//         return res.status(400).json(`❌ Fallo en getPadelMatchByAuthor: ${error.message}`);
+//     }
+// };
 
 const updatePadelMatch = async (req, res, next) => {
     try {
@@ -62,8 +77,23 @@ const updatePadelMatch = async (req, res, next) => {
 const deletePadelMatch = async (req, res, next) => {
     try {
         const { id } = req.params;
+
+        const findPadelMatch = await PadelMatch.findById(id);
+        if (!findPadelMatch) {
+            return res.status(400).json({ message: "No existe ese partido." });
+        }
+
+        const authorPadelMatch = await User.findById(req.user);
+        const authorIDPadelMatch = findPadelMatch.author.toString();
+
+        const userChecked = idAndRoleChecked(authorIDPadelMatch, authorPadelMatch);
+        if (userChecked) {
+            return res.status(400).json({ message: userChecked });
+        }
+
         const padelMatchDeleted = await PadelMatch.findByIdAndDelete(id);
-        padelMatchDeleted ? res.status(200).json({ message: `Partido eliminado.`, padelMatchDeleted }) : res.status(400).json({ message: `No se ha encontrado ese partido.` });
+        deleteImage(padelMatchDeleted.image);
+        resultPadelMatchDeleted(res, padelMatchDeleted);
     } catch (error) {
         return res.status(400).json(`❌ Fallo en deletePadelMatch: ${error.message}`);
     }
@@ -72,6 +102,7 @@ const deletePadelMatch = async (req, res, next) => {
 module.exports = {
     createPadelMatch,
     getPadelMatches,
-    getPadelMatchByDate,
+    getPadelMatchByDay,
+    updatePadelMatch,
     deletePadelMatch
 };
